repair bugs

+++ src/main/java/org/temkarus0070/memory/MemoryTest.java
@@ -6,5 +6,5 @@ import javax.xml.parsers.ParserConfigurationException;
 
 public interface MemoryTest {
     public void doTest(InputStream file, Set<String> searchedValues, String tagToCalcCount) throws ParserConfigurationException;
-    public double getTestResult() throws IllegalArgumentException;
+    public long getTestResult() throws IllegalArgumentException;
 }
